% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_draw.R
\name{leg_draw}
\alias{leg_draw}
\title{Draw a map legend}
\usage{
leg_draw(
  x,
  pos = "bottomright",
  size = 1,
  bg = "white",
  fg = "black",
  frame = TRUE,
  mar = par("mar")
)
}
\arguments{
\item{x}{legend}

\item{pos}{pos}

\item{size}{size}

\item{bg}{bg}

\item{fg}{fg}

\item{frame}{frame}

\item{mar}{mar}
}
\description{
Draw a map legend with several elements.
}
\examples{
plot(1:10)
}
